move files to sync module
