+++ Tests/MongoSwiftSyncTests/MongoClientTests.swift
@@ -1,5 +1,5 @@
 import Foundation
-@testable import MongoSwift
+import MongoSwift
 import Nimble
 import TestsCommon
 import XCTest
