--- Tests/MongoSwiftSyncTests/Placeholder.swift
@@ -1,1 +1,0 @@
-// placeholder file to satisfy SwiftPM
