+++ Tests/MongoSwiftSyncTests/SDAMMonitoringTests.swift
@@ -80,8 +80,8 @@ final class SDAMTests: MongoSwiftTestCase {
 
         let event1 = receivedEvents[1] as! TopologyDescriptionChangedEvent
         expect(event1.topologyId).to(equal(event0.topologyId))
-        expect(event1.previousDescription.type).to(equal(TopologyDescription.TopologyType.unknown))
-        expect(event1.newDescription.type).to(equal(TopologyDescription.TopologyType.single))
+        expect(event1.previousDescription.type).to(equal(.unknown))
+        expect(event1.newDescription.type).to(equal(.single))
         // This is a bit of a deviation from the SDAM spec tests linked above. However, this is how mongoc responds so
         // there is no other way to get around this.
         expect(event1.newDescription.servers).to(beEmpty())
@@ -100,18 +100,18 @@ final class SDAMTests: MongoSwiftTestCase {
         let newServer = event3.newDescription
         self.checkDefaultHostPort(newServer, hostlist)
         self.checkEmptyLists(newServer)
-        expect(newServer.type).to(equal(ServerDescription.ServerType.standalone))
+        expect(newServer.type).to(equal(.standalone))
 
         let event4 = receivedEvents[4] as! TopologyDescriptionChangedEvent
         expect(event4.topologyId).to(equal(event3.topologyId))
         let prevTopology = event4.previousDescription
-        expect(prevTopology.type).to(equal(TopologyDescription.TopologyType.single))
+        expect(prevTopology.type).to(equal(.single))
         expect(prevTopology.servers).to(beEmpty())
 
         let newTopology = event4.newDescription
-        expect(newTopology.type).to(equal(TopologyDescription.TopologyType.single))
+        expect(newTopology.type).to(equal(.single))
         self.checkDefaultHostPort(newTopology.servers[0], hostlist)
-        expect(newTopology.servers[0].type).to(equal(ServerDescription.ServerType.standalone))
+        expect(newTopology.servers[0].type).to(equal(.standalone))
         self.checkEmptyLists(newTopology.servers[0])
     }
 }
